extract maven repo interaction logic for JVM ecosystems reuse
